pyproject.toml
[project]
name = "cybersecurity-rss"
version = "0.2.0"
description = "Multi-board RSS aggregator with LLM-curated Chinese summaries and GitHub Pages output."
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "feedparser>=6.0.11",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.3",
]

[tool.ruff]
line-length = 100
target-version = "py311"
